Added migration for creating new password_securities table and updated password_expiry_days column in the site table to a tinyInteger in the AddPasswordExpiryDaysToSiteTable migration

+++ database/migrations/2020_10_21_155839_add_password_expiry_days_to_site_table.php
@@ -13,7 +13,7 @@ class AddPasswordExpiryDaysToSiteTable extends Migration
     public function up()
     {
         Schema::table('site', function (Blueprint $table) {
-            $table->integer('password_expiry_days')->after('type')->nullable()->default(90);
+            $table->tinyInteger('password_expiry_days')->after('type')->nullable()->default(90);
         });
     }
